requirements.txt
torch
pandas
numpy
PyYAML
matplotlib
seaborn
scipy
tqdm
torch_geometric     